// GMAOWB/src/app/Models/Intervention.ts
export class Intervention {
   NOInter: number;
   NumDem: number;
   Matricule:string;
   Intervenant: string;
   Categorie:string;
   TauxHoraire: number;
   DDebutEstim: Date;
   HDebutEstim: Date;
   DFinEstim: Date;
    HFinEstim: Date;
    DDebut: Date;
    HDebut:Date;
    DFin:Date;
    HFin:Date;
    NHeure:Date;
    DateSaisie:Date;
    HeureSaisie:Date;
}